+++ src/main/ruby/truffleruby/core/match_data.rb
@@ -137,6 +137,19 @@ class MatchData
     out
   end
 
+  def match(n)
+    # Similar, but #match accepts only single index/name, but not a range or an optional length.
+    number = Truffle::Type.rb_check_convert_type(n, Integer, :to_int)
+    return self[number] if number
+    # To convert the last type (String) we used rb_convert_type instead of rb_check_convert_type which throws an exception
+    name = Truffle::Type.rb_check_convert_type(n, Symbol, :to_sym) || Truffle::Type.rb_convert_type(n, String, :to_str)
+    self[name]
+  end
+
+  def match_length(n)
+    match(n)&.length
+  end
+
   def to_s
     self[0]
   end